refactor: refactor admin endpoints into smaller files
