--- libs/core-server/src/__tests__/setup.js
@@ -1,5 +1,0 @@
-global.console = {
-  log: jest.fn(),
-  debug: jest.fn(),
-  info: jest.fn(),
-};
